requirements.txt
matplotlib
pandas
pympler
python-Levenshtein